apps/ingestion/backfill/requirements.txt
dlt
kafka-python
psycopg2-binary
pandas
python-dotenv
requests
